Move types/values for errors and query params out of routes
